backend/employee/requirements.txt
psycopg[binary]
djangorestframework-simplejwt[blacklist]
djangorestframework
django-cors-headers
gunicorn
whitenoise
python-dotenv
reportlab
django-ratelimit
django-csp
argon2-cffi
bleach
django-environ

# Caching
redis
django-redis


# Email
django-anymail

# Testing
pytest
pytest-django
pytest-cov
factory-boy
faker

# Development
django-debug-toolbar
ipython